add sparse pattern discovery for dmat so we don't have to do dense assembly - we precompute the sparsity pattern  (coupling between block j and j-1) until it stabilizes (usually at n=D+1) - then we use this pattern to map into non-zero indices for arbitrary n. also add csc storage for it, and testing. next step is do the same with kmat, jmat.

+++ include/jkmat.hh
@@ -136,7 +136,7 @@ struct KMat
     // K = Vtgt^T * diag(wq) * Vsrc
     // K[i,j] = sum_p Vtgt[p+i*npts] * wq[p] * Vsrc[p+j*npts]
     constexpr Real KMAT_PRUNE_TOL =
-      Real(100) * std::numeric_limits<Real>::epsilon();
+      Real(1000) * std::numeric_limits<Real>::epsilon();
     
     #pragma omp parallel for schedule(static)
     for (int i = 0; i < M; ++i)
